+++ newsAnalysis/datastore/model.py
